+++ readme.txt
@@ -1,1 +1,46 @@
-hector branch is being handled by habi8
+# Gomoku AI with Minimax and Alpha-Beta Pruning
+
+![Gomoku Game Screenshot](./images.ss1.png) 
+![Gomoku Game Screenshot](./images.ss2.png) 
+
+A Python implementation of the classic Gomoku (Five in a Row) game with an AI opponent using the Minimax algorithm and Alpha-Beta pruning optimization. The game features a Pygame-based graphical interface.
+
+## Features
+
+- Classic Gomoku (5-in-a-row) gameplay
+- Play Against Human opponent or AI opponent
+- Minimax algorithm with Alpha-Beta pruning for efficient decision making
+- Pygame-based graphical interface
+- Win detection and game state tracking
+
+## Requirements
+
+- Python 3.7+
+- Pygame
+
+## Installation & Running Instructions
+
+1. **Clone the repository**:
+   ```bash
+   git clone https://github.com/sejutiii/Gomoku
+   cd Gomoku
+
+2. **Set up a virtual environment (recommended)**:
+    ```bash
+    python -m venv venv
+
+   On Windows- 
+    ```bash
+    .\venv\Scripts\activate
+
+   On Linux/ MacOS- 
+    ```bash
+    source venv/bin/activate
+
+3. **Install dependencies:**
+    ```bash
+    pip install -r requirements.txt
+
+4. **Run The Game:**
+    ```bash
+    python Gomoku.py
